[FEATURE] Using oneSignal push notification to offline user

See merge request 9thflr.rd/api-chatshier!377

+++ helpers/onesignal.js
@@ -1,66 +1,36 @@
 // https://documentation.onesignal.com/reference
 module.exports = (function() {
-    const oneSignalCfg = require('../config/onesignal');
+    const CHATSHIER_CFG = require('../config/chatshier');
     const request = require('request');
+
+    const API_ENDPOINT = 'https://onesignal.com/api/v1';
+    const APIKEY_NOT_FOUND = new Error('APIKEY_NOT_FOUND');
+    const APIKEY_TABLE = CHATSHIER_CFG.ONESIGNAL_APIKEYS;
 
     const sendRequest = (options) => {
         return new Promise((resolve, reject) => {
             request(options, (error, res, body) => {
+                let canParseJSON =
+                    (res.headers['Content-Type'] && res.headers['Content-Type'].includes('application/json')) ||
+                    (res.headers['content-type'] && res.headers['content-type'].includes('application/json')) ||
+                    ('string' === typeof body && body.length > 0 &&
+                    ((body.startsWith('{') && body.endsWith('}')) || (body.startsWith('[') && body.endsWith(']'))));
+
+                body = canParseJSON && 'string' === typeof body ? JSON.parse(body) : body;
+
                 if (error || res.statusCode >= 300) {
                     return reject(body);
                 }
-
-                let hasJsonHeader =
-                    (res.headers['Content-Type'] && res.headers['Content-Type'].includes('application/json')) ||
-                    (res.headers['content-type'] && res.headers['content-type'].includes('application/json'));
-                resolve(hasJsonHeader && 'string' === typeof body ? JSON.parse(body) : body);
+                resolve(body);
             });
         });
     };
 
     class OneSignalHelper {
         constructor() {
-            this.API_ENDPOINT = oneSignalCfg.API_ENDPOINT;
-            this.APPID = oneSignalCfg.APPID;
-            this.REST_API_KEY = oneSignalCfg.REST_API_KEY;
-            this.USER_AUTH_KEY = oneSignalCfg.USER_AUTH_KEY;
             this.headers = {
                 'Content-Type': 'application/json; charset=utf-8'
             };
-        }
-
-        /** @private */
-        _headersBuilder(headers) {
-            return Object.assign(headers, this.headers);
-        }
-
-        /**
-         * @returns {Promise<OneSignal.App[]>}
-         */
-        viewApps() {
-            let options = {
-                method: 'GET',
-                url: this.API_ENDPOINT + '/apps',
-                headers: this._headersBuilder({
-                    Authorization: 'Basic ' + this.USER_AUTH_KEY
-                })
-            };
-            return sendRequest(options);
-        }
-
-        /**
-         * @param {string} appId
-         * @returns {Promise<OneSignal.App>}
-         */
-        viewApp(appId) {
-            let options = {
-                method: 'GET',
-                url: this.API_ENDPOINT + '/apps/' + appId,
-                headers: this._headersBuilder({
-                    Authorization: 'Basic ' + this.USER_AUTH_KEY
-                })
-            };
-            return sendRequest(options);
         }
 
         /**
@@ -69,30 +39,17 @@ module.exports = (function() {
          * @param {number} [offset=0]
          * @returns {Promise<OneSignal.Devices>}
          */
-        viewDevices(appId, limit, offset) {
-            limit = limit || 300;
-            offset = !offset ? 0 : offset;
+        viewDevices(appId, limit = 300, offset = 0) {
+            let restApiKey = APIKEY_TABLE[appId];
+            if (!restApiKey) {
+                return Promise.reject(APIKEY_NOT_FOUND);
+            }
 
             let options = {
                 method: 'GET',
-                url: this.API_ENDPOINT + '/players?app_id=' + appId + '&limit=' + limit + '&offset=' + offset,
-                headers: this._headersBuilder({
-                    Authorization: 'Basic ' + this.REST_API_KEY
-                })
-            };
-            return sendRequest(options);
-        }
-
-        /**
-         * @param {string} playerId
-         * @returns {Promise<OneSignal.Device>}
-         */
-        viewDevice(playerId) {
-            let options = {
-                method: 'GET',
-                url: this.API_ENDPOINT + '/players/' + playerId,
-                headers: this._headersBuilder({
-                    Authorization: 'Basic ' + this.REST_API_KEY
+                url: API_ENDPOINT + '/players?app_id=' + appId + '&limit=' + limit + '&offset=' + offset,
+                headers: Object.assign(this.headers, {
+                    Authorization: 'Basic ' + restApiKey
                 })
             };
             return sendRequest(options);
@@ -100,19 +57,42 @@ module.exports = (function() {
 
         /**
          * @param {string} appId
-         * @param {number} [limit=300]
-         * @param {number} [offset=0]
-         * @returns {Promise<OneSignal.Notifications>}
+         * @param {string} playerId
+         * @returns {Promise<OneSignal.Device>}
          */
-        viewNotifications(appId, limit, offset) {
-            limit = limit || 50;
-            offset = !offset ? 0 : offset;
+        viewDevice(appId, playerId) {
+            let restApiKey = APIKEY_TABLE[appId];
+            if (!restApiKey) {
+                return Promise.reject(APIKEY_NOT_FOUND);
+            }
 
             let options = {
                 method: 'GET',
-                url: this.API_ENDPOINT + '/notifications?app_id=' + appId + '&limit=' + limit + '&offset=' + offset,
-                headers: this._headersBuilder({
-                    Authorization: 'Basic ' + this.REST_API_KEY
+                url: API_ENDPOINT + '/players/' + playerId,
+                headers: Object.assign(this.headers, {
+                    Authorization: 'Basic ' + restApiKey
+                })
+            };
+            return sendRequest(options);
+        }
+
+        /**
+         * @param {string} appId
+         * @param {number} [limit=50]
+         * @param {number} [offset=0]
+         * @returns {Promise<OneSignal.Notifications>}
+         */
+        viewNotifications(appId, limit = 50, offset = 0) {
+            let restApiKey = APIKEY_TABLE[appId];
+            if (!restApiKey) {
+                return Promise.reject(APIKEY_NOT_FOUND);
+            }
+
+            let options = {
+                method: 'GET',
+                url: API_ENDPOINT + '/notifications?app_id=' + appId + '&limit=' + limit + '&offset=' + offset,
+                headers: Object.assign(this.headers, {
+                    Authorization: 'Basic ' + restApiKey
                 })
             };
             return sendRequest(options);
@@ -124,11 +104,16 @@ module.exports = (function() {
          * @returns {Promise<OneSignal.Notification>}
          */
         viewNotification(appId, notificationId) {
+            let restApiKey = APIKEY_TABLE[appId];
+            if (!restApiKey) {
+                return Promise.reject(APIKEY_NOT_FOUND);
+            }
+
             let options = {
                 method: 'GET',
-                url: this.API_ENDPOINT + '/notifications/' + notificationId + '?app_id=' + appId,
-                headers: this._headersBuilder({
-                    Authorization: 'Basic ' + this.REST_API_KEY
+                url: API_ENDPOINT + '/notifications/' + notificationId + '?app_id=' + appId,
+                headers: Object.assign(this.headers, {
+                    Authorization: 'Basic ' + restApiKey
                 })
             };
             return sendRequest(options);
@@ -140,11 +125,16 @@ module.exports = (function() {
          * @returns {Promise<{ id: string, recipients: number, errors?: any }>}
          */
         createNotification(appId, notification) {
+            let restApiKey = APIKEY_TABLE[appId];
+            if (!restApiKey) {
+                return Promise.reject(APIKEY_NOT_FOUND);
+            }
+
             let options = {
                 method: 'POST',
-                url: this.API_ENDPOINT + '/notifications/?app_id=' + appId,
-                headers: this._headersBuilder({
-                    Authorization: 'Basic ' + this.REST_API_KEY
+                url: API_ENDPOINT + '/notifications/?app_id=' + appId,
+                headers: Object.assign(this.headers, {
+                    Authorization: 'Basic ' + restApiKey
                 }),
                 json: notification
             };
@@ -157,29 +147,21 @@ module.exports = (function() {
          * @returns {Promise<{ success: 'true' }>}
          */
         cancelNotification(appId, notificationId) {
+            let restApiKey = APIKEY_TABLE[appId];
+            if (!restApiKey) {
+                return Promise.reject(APIKEY_NOT_FOUND);
+            }
+
             let options = {
                 method: 'DELETE',
-                url: this.API_ENDPOINT + '/notifications/' + notificationId + '?app_id=' + appId,
-                headers: this._headersBuilder({
-                    Authorization: 'Basic ' + this.REST_API_KEY
+                url: API_ENDPOINT + '/notifications/' + notificationId + '?app_id=' + appId,
+                headers: Object.assign(this.headers, {
+                    Authorization: 'Basic ' + restApiKey
                 })
             };
             return sendRequest(options);
         }
     }
 
-    let instance = new OneSignalHelper();
-
-    // /** @type {OneSignal.Notification} */
-    // let notification = {
-    //     included_segments: ['Active Users'],
-    //     contents: {
-    //         en: 'Test notification'
-    //     }
-    // };
-    // instance.createNotification(instance.APPID, notification).then((resJson) => {
-    //     console.log(resJson);
-    // });
-
-    return instance;
+    return new OneSignalHelper();
 })();
